% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{createDenTable}
\alias{createDenTable}
\title{Based on createDenData create a gtable}
\usage{
createDenTable(data, style, label = TRUE, transform = NULL)
}
\arguments{
\item{data}{Data as returned by \code{\link{createDenData}}}

\item{style}{A complete ggplot theme}

\item{label}{Logical should the sets be labelled}

\item{transform}{A trans object or the name of one from scales}
}
\value{
A gtable object ready to draw
}
\description{
Based on createDenData create a gtable
}

