% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{createIceTable}
\alias{createIceTable}
\title{Based on createIceData create a gtable}
\usage{
createIceTable(data, style, label = TRUE, yaxis = "left",
  transform = NULL, showHierarchy = FALSE)
}
\arguments{
\item{data}{Data as returned by \code{\link{createIceData}}}

\item{style}{A complete ggplot theme}

\item{label}{Logical should the sets be labelled}

\item{yaxis}{The position of the y-axis}

\item{transform}{A trans object or the name of one from scales}

\item{showHierarchy}{Should a dendrogram be plotted on top}
}
\value{
A gtable object ready to draw
}
\description{
Based on createIceData create a gtable
}

