% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hSet.R
\name{format_sets}
\alias{format_sets}
\title{Parse different set formats into ngCMatrix format}
\usage{
format_sets(x)
}
\arguments{
\item{x}{The sets to be formatted}
}
\value{
A presence/absence matrix with sets as columns and elements as rows,
formatted as a ngCMatrix
}
\description{
This function is intended to ensure proper formatting of the supplied sets to
the \code{\link{create_hierarchy}} constructor. Support for other input
objects can be created by writting a format_sets method for the class.
}
\examples{
data('twitter')
format_sets(as.matrix(twitter))

}
\keyword{internal}

