% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\docType{data}
\name{twitter}
\alias{twitter}
\title{Followers of 100 twitter users}
\format{
A ngCMatrix with the sets (users) as columns and elements (followers) as rows
}
\source{
\url{https://snap.stanford.edu/data/}
}
\usage{
twitter
}
\description{
This dataset captures the followers of 100 highly followed twitter users. The
dataset is anonymized and based on the ego-Twitter network from Stanford
Large Network Dataset Collection.
}
\keyword{datasets}
